Rename `relay/client.rs` to `relay/conn.rs`
