+++ iroh-net/src/magicsock/relay_actor.rs
@@ -20,7 +20,7 @@ use tracing::{debug, info, info_span, trace, warn, Instrument};
 
 use crate::{
     key::{PublicKey, PUBLIC_KEY_LENGTH},
-    relay::{self, client::ReceivedMessage, http::ClientError, RelayUrl, MAX_PACKET_SIZE},
+    relay::{self, conn::ReceivedMessage, http::ClientError, RelayUrl, MAX_PACKET_SIZE},
 };
 
 use super::{ActorMessage, MagicSock};
@@ -214,7 +214,7 @@ impl ActiveRelay {
                 }
 
                 match msg {
-                    relay::client::ReceivedMessage::ReceivedPacket { source, data } => {
+                    relay::conn::ReceivedMessage::ReceivedPacket { source, data } => {
                         trace!(len=%data.len(), "received msg");
                         // If this is a new sender we hadn't seen before, remember it and
                         // register a route for this peer.
@@ -244,7 +244,7 @@ impl ActiveRelay {
 
                         ReadResult::Continue
                     }
-                    relay::client::ReceivedMessage::Ping(data) => {
+                    relay::conn::ReceivedMessage::Ping(data) => {
                         // Best effort reply to the ping.
                         let dc = self.relay_client.clone();
                         tokio::task::spawn(async move {
@@ -254,8 +254,8 @@ impl ActiveRelay {
                         });
                         ReadResult::Continue
                     }
-                    relay::client::ReceivedMessage::Health { .. } => ReadResult::Continue,
-                    relay::client::ReceivedMessage::PeerGone(key) => {
+                    relay::conn::ReceivedMessage::Health { .. } => ReadResult::Continue,
+                    relay::conn::ReceivedMessage::PeerGone(key) => {
                         self.relay_routes.retain(|peer| peer != &key);
                         ReadResult::Continue
                     }
