+++ iroh-net/src/relay/server/client_conn.rs
@@ -555,7 +555,7 @@ mod tests {
         // send packet
         println!("  send packet");
         let data = b"hello world!";
-        crate::relay::client::send_packet(&mut io_rw, &None, target, Bytes::from_static(data))
+        crate::relay::conn::send_packet(&mut io_rw, &None, target, Bytes::from_static(data))
             .await?;
         let msg = server_channel_r.recv().await.unwrap();
         match msg {
@@ -575,7 +575,7 @@ mod tests {
         let mut disco_data = crate::disco::MAGIC.as_bytes().to_vec();
         disco_data.extend_from_slice(target.as_bytes());
         disco_data.extend_from_slice(data);
-        crate::relay::client::send_packet(&mut io_rw, &None, target, disco_data.clone().into())
+        crate::relay::conn::send_packet(&mut io_rw, &None, target, disco_data.clone().into())
             .await?;
         let msg = server_channel_r.recv().await.unwrap();
         match msg {
@@ -630,7 +630,7 @@ mod tests {
         let data = b"hello world!";
         let target = SecretKey::generate().public();
 
-        crate::relay::client::send_packet(&mut io_rw, &None, target, Bytes::from_static(data))
+        crate::relay::conn::send_packet(&mut io_rw, &None, target, Bytes::from_static(data))
             .await?;
         let msg = server_channel_r.recv().await.unwrap();
         match msg {
